WinAdapter: Export helper functions that operate on support types
The implementation of these support types completely depends on DXC.  As
such applications using libdxcompiler.so have to guess and/or look at
the source code to understand how they are implemented (eg prefix size
for BSTR), and assume this is not going to change.  Applications using
these newly exported symbols can do it "the right way" as if they are on
Windows.